use temp tables for test-only models. this requires some hackish explicit requiring of code and namespace checking to work around rails app load order. I tried using Temping gem but found that psql temporary tables were getting removed at unpredictable points when running full suite (probably because connections were open/closed)
