--- db/migrate/20151210225310_create_test_client_requests.rb
@@ -1,10 +1,0 @@
-class CreateTestClientRequests < ActiveRecord::Migration
-  def change
-    create_table :test_client_requests do |t|
-      t.decimal :amount
-      t.string :project_title
-
-      t.timestamps null: false
-    end
-  end
-end
